requirements.txt
requests
pandas
numpy
tqdm
scikit-learn
streamlit
python-dotenv